refactor: extract target formatters
